requirements.txt
Flask>=3.0,<4.0
groq>=0.4.0
numpy>=1.23
scikit-learn>=1.3
python-dotenv>=1.0
